+++ backend/src/main/java/com/kozimor/wms/Database/Model/OrderLine.java
@@ -33,7 +33,7 @@ public class OrderLine {
 
     @Min(1)
     @Column(name = "quantity", nullable = false)
-    private Integer quantity;
+    private Double quantity;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "transaction_id")
